occupants

+++ database/migrations/2024_06_09_194528_create_zonecoords_table.php
@@ -13,16 +13,13 @@ return new class extends Migration
     {
         Schema::create('zonecoords', function (Blueprint $table) {
             $table->id();
-            $table->decimal('latitude');
-            $table->decimal('longitude');
+            $table->double('latitude');
+            $table->double('longitude');
             $table->foreignId('zone_id')->constrained()->onDelete('cascade');
             $table->timestamps();
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('zonecoords');
